Evaluate
Added Evaluate class that replicates our evaluation logic. Due to the challenge rules, my evaluation function is required to be in MyBot. However, it would be cleaner if I could call a separate class. That would allow me to update the logic and impact both MyBot as well as the BoardAnalysis. For now, we replicate.

I also created a simple test to input a FEN and output a score for each element. I tested based on a few famous games.