Board Analysis
Added a previous chess engine that I built in Python that has the ability to read PGN games. I extended to be able to get information about the game as well as a random move and the FEN for the random move. We can use this to analyze many games.